requirements.txt
Flask
flask-cors
tensorflow
python-dotenv
pydantic